// src/api/models/Card.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamBoard API
 * Kanban Board API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
    UserToJSONTyped,
} from './User';

/**
 * 
 * @export
 * @interface Card
 */
export interface Card {
    /**
     * 
     * @type {number}
     * @memberof Card
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Card
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof Card
     */
    description?: string;
    /**
     * 
     * @type {User}
     * @memberof Card
     */
    assignedTo?: User;
    /**
     * 
     * @type {number}
     * @memberof Card
     */
    columnId?: number;
}

/**
 * Check if a given object implements the Card interface.
 */
export function instanceOfCard(value: object): value is Card {
    return true;
}

export function CardFromJSON(json: any): Card {
    return CardFromJSONTyped(json, false);
}

export function CardFromJSONTyped(json: any, ignoreDiscriminator: boolean): Card {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'assignedTo': json['assignedTo'] == null ? undefined : UserFromJSON(json['assignedTo']),
        'columnId': json['columnId'] == null ? undefined : json['columnId'],
    };
}

export function CardToJSON(json: any): Card {
    return CardToJSONTyped(json, false);
}

export function CardToJSONTyped(value?: Card | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'assignedTo': UserToJSON(value['assignedTo']),
        'columnId': value['columnId'],
    };
}