// src/api/models/Column.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamBoard API
 * Kanban Board API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Card } from './Card';
import {
    CardFromJSON,
    CardFromJSONTyped,
    CardToJSON,
    CardToJSONTyped,
} from './Card';

/**
 * 
 * @export
 * @interface Column
 */
export interface Column {
    /**
     * 
     * @type {number}
     * @memberof Column
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Column
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof Column
     */
    position?: number;
    /**
     * 
     * @type {Array<Card>}
     * @memberof Column
     */
    cards?: Array<Card>;
    /**
     * 
     * @type {number}
     * @memberof Column
     */
    projectId?: number;
}

/**
 * Check if a given object implements the Column interface.
 */
export function instanceOfColumn(value: object): value is Column {
    return true;
}

export function ColumnFromJSON(json: any): Column {
    return ColumnFromJSONTyped(json, false);
}

export function ColumnFromJSONTyped(json: any, ignoreDiscriminator: boolean): Column {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'position': json['position'] == null ? undefined : json['position'],
        'cards': json['cards'] == null ? undefined : ((json['cards'] as Array<any>).map(CardFromJSON)),
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
    };
}

export function ColumnToJSON(json: any): Column {
    return ColumnToJSONTyped(json, false);
}

export function ColumnToJSONTyped(value?: Column | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'position': value['position'],
        'cards': value['cards'] == null ? undefined : ((value['cards'] as Array<any>).map(CardToJSON)),
        'projectId': value['projectId'],
    };
}